+++ threedi_schematisation_editor/processing/alghorithms_inflow.py
@@ -14,18 +14,17 @@ from qgis.core import (
     QgsProcessingParameterVectorLayer,
     QgsProject,
 )
-from qgis.PyQt.QtCore import QCoreApplication
 
 from threedi_schematisation_editor.enumerators import SewerageType
 from threedi_schematisation_editor.utils import get_feature_by_id, get_next_feature_id, spatial_index
 
 
-class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
-    """Link (DWF) surfaces to connection nodes."""
+class LinkToConnectionNodesAlgorithm(QgsProcessingAlgorithm):
+    """Base algorithm for linking surface or DWF polygons to connection nodes."""
 
-    SURFACE_LAYER = "SURFACE_LAYER"
-    SELECTED_SURFACES = "SELECTED_SURFACES"
-    SURFACE_MAP_LAYER = "SURFACE_MAP_LAYER"
+    POLYGON_LAYER = "POLYGON_LAYER"
+    SELECTED_POLYGONS = "SELECTED_POLYGONS"
+    MAP_LAYER = "MAP_LAYER"
     PIPE_LAYER = "PIPE_LAYER"
     SELECTED_PIPES = "SELECTED_PIPES"
     NODE_LAYER = "NODE_LAYER"
@@ -34,35 +33,38 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
     SANITARY_SEWER_PREFERENCE = "SANITARY_SEWER_PREFERENCE"
     SEARCH_DISTANCE = "SEARCH_DISTANCE"
 
-    def tr(self, string):
-        return QCoreApplication.translate("Processing", string)
+    def parameter_names(self):
+        # Not implemented as proper abstractmethod because QgsProcessingAlgorithm already has a metaclass
+        # and setting ABCMeta as metaclass creates complicated problems
+        return NotImplementedError("Subclasses must implement this method")
 
-    def createInstance(self):
-        return LinkSurfacesWithNodes()
+    def default_values(self):
+        # Not implemented as proper abstractmethod because QgsProcessingAlgorithm already has a metaclass
+        # and setting ABCMeta as metaclass creates complicated problems
+        return NotImplementedError("Subclasses must implement this method")
 
-    def name(self):
-        return "threedi_map_surfaces_to_connection_nodes"
-
-    def displayName(self):
-        return self.tr("Map surfaces to connection nodes")
+    def polygon_id_field(self):
+        # Not implemented as proper abstractmethod because QgsProcessingAlgorithm already has a metaclass
+        # and setting ABCMeta as metaclass creates complicated problems
+        return NotImplementedError("Subclasses must implement this method")
 
     def group(self):
-        return self.tr("Inflow")
+        return "Inflow"
 
     def groupId(self):
         return "0d"
 
     def shortHelpString(self):
-        return self.tr(
-            """
-            <p>Connect surfaces to the sewer system by creating surface map features. The new features are added to the surface layer directly.</p>
-            <p>For each surface, the nearest pipe is found; the surface is mapped to the the nearest of this pipe's connection nodes.</p>
-            <p>In some cases, you may want to prefer e.g. stormwater drains over combined sewers. This can be done by setting the stormwater sewer preference to a value greater than zero.</p>
+        return (
+            f"""
+            <p>Connect {self.parameter_names[self.POLYGON_LAYER]} features to the sewer system by creating {self.default_values[self.MAP_LAYER]} features. The new features are added to the {self.parameter_names[self.MAP_LAYER]} directly.</p>
+            <p>For each {self.parameter_names[self.POLYGON_LAYER]} feature, the nearest pipe is found; it is mapped to the nearest of this pipe's connection nodes.</p>
+            <p>In some cases, you may want to prefer e.g. stormwater drains or sanitary sewers over combined sewers. This can be done by setting the stormwater sewer preference to a value greater than zero.</p>
             <h3>Parameters</h3>
-            <h4>(DWF) surface layer</h4>
-            <p>Surface or DWF surface layer that is added to the project with the 3Di Schematisation Editor.</p>
-            <h4>(DWF) surface map layer</h4>
-            <p>Surface map or DWF surface map layer that is added to the project with the 3Di Schematisation Editor.</p>
+            <h4>{self.parameter_names[self.POLYGON_LAYER]}</h4>
+            <p>{self.parameter_names[self.POLYGON_LAYER]} that is added to the project with the 3Di Schematisation Editor.</p>
+            <h4>{self.parameter_names[self.MAP_LAYER]}</h4>
+            <p>{self.parameter_names[self.MAP_LAYER]} that is added to the project with the 3Di Schematisation Editor.</p>
             <h4>Pipe layer</h4>
             <p>Pipe layer that is added to the project with the 3Di Schematisation Editor.</p>
             <h4>Connection node layer</h4>
@@ -70,9 +72,9 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
             <h4>Sewerage types</h4>
             <p>Only pipes of the selected sewerage types will be used in the algorithm</p>
             <h4>Stormwater sewer preference</h4>
-            <p>This value (in meters) will be subtracted from the distance between the surface and the stormwater drain. For example: there is a combined sewer within 10 meters from the surface, and a stormwater drain within 11 meters; if the stormwater sewer preference is 2 m, the algorithm will use 11 - 2 = 9 m as distance to the stormwater sewer, so the surface will be mapped to one of the stormwater drain's connection nodes, instead of to the combined sewer's connection nodes.</p>
+            <p>This value (in meters) will be subtracted from the distance between the {self.default_values[self.POLYGON_LAYER]} and the stormwater drain. For example: there is a combined sewer within 10 meters from the {self.default_values[self.POLYGON_LAYER]}, and a stormwater drain within 11 meters; if the stormwater sewer preference is 2 m, the algorithm will use 11 - 2 = 9 m as distance to the stormwater sewer, so the {self.default_values[self.POLYGON_LAYER]} will be mapped to one of the stormwater drain's connection nodes, instead of to the combined sewer's connection nodes.</p>
             <h4>Sanitary sewer preference</h4>
-            <p>This value (in meters) will be subtracted from the distance between the surface and the sanitary sewer. See 'stormwater sewer preference' for further explanation.</p>
+            <p>This value (in meters) will be subtracted from the distance between the {self.default_values[self.POLYGON_LAYER]} and the sanitary sewer. See 'stormwater sewer preference' for further explanation.</p>
             <h4>Search distance</h4>
             <p>Only pipes within search distance (m) from the surface will be used in the algorithm.</p>
             """
@@ -81,30 +83,30 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
     def initAlgorithm(self, config=None):
         self.addParameter(
             QgsProcessingParameterVectorLayer(
-                self.SURFACE_LAYER,
-                self.tr("Surface layer"),
+                self.POLYGON_LAYER,
+                self.parameter_names[self.POLYGON_LAYER],
                 [QgsProcessing.TypeVectorPolygon],
-                defaultValue="Surface",
+                defaultValue=self.default_values[self.POLYGON_LAYER],
             )
         )
         self.addParameter(
             QgsProcessingParameterBoolean(
-                self.SELECTED_SURFACES,
-                self.tr("Selected surfaces only"),
+                self.SELECTED_POLYGONS,
+                self.parameter_names[self.SELECTED_POLYGONS],
             )
         )
         self.addParameter(
             QgsProcessingParameterVectorLayer(
-                self.SURFACE_MAP_LAYER,
-                self.tr("Surface map layer"),
+                self.MAP_LAYER,
+                self.parameter_names[self.MAP_LAYER],
                 [QgsProcessing.TypeVectorLine],
-                defaultValue="Surface map",
+                defaultValue=self.default_values[self.MAP_LAYER],
             )
         )
         self.addParameter(
             QgsProcessingParameterVectorLayer(
                 self.PIPE_LAYER,
-                self.tr("Pipe layer"),
+                "Pipe layer",
                 [QgsProcessing.TypeVectorLine],
                 defaultValue="Pipe",
             )
@@ -112,13 +114,13 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         self.addParameter(
             QgsProcessingParameterBoolean(
                 self.SELECTED_PIPES,
-                self.tr("Selected pipes only"),
+                "Selected pipes only",
             )
         )
         self.addParameter(
             QgsProcessingParameterVectorLayer(
                 self.NODE_LAYER,
-                self.tr("Connection node layer"),
+                "Connection node layer",
                 [QgsProcessing.TypeVectorPoint],
                 defaultValue="Connection node",
             )
@@ -126,15 +128,15 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         self.addParameter(
             QgsProcessingParameterEnum(
                 self.SEWERAGE_TYPES,
-                self.tr("Sewerage types"),
+                "Sewerage types",
                 allowMultiple=True,
                 options=[e.name for e in SewerageType],
-                defaultValue=[SewerageType.COMBINED_SEWER.value, SewerageType.STORM_DRAIN.value],
+                defaultValue=self.default_values[self.SEWERAGE_TYPES],
             )
         )
         storm_pref = QgsProcessingParameterNumber(
             self.STORMWATER_SEWER_PREFERENCE,
-            self.tr("Stormwater sewer preference [m]"),
+            "Stormwater sewer preference [m]",
             type=QgsProcessingParameterNumber.Double,
             defaultValue=0.0,
         )
@@ -143,7 +145,7 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         self.addParameter(storm_pref)
         sanitary_pref = QgsProcessingParameterNumber(
             self.SANITARY_SEWER_PREFERENCE,
-            self.tr("Sanitary sewer preference [m]"),
+            "Sanitary sewer preference [m]",
             type=QgsProcessingParameterNumber.Double,
             defaultValue=0.0,
         )
@@ -152,7 +154,7 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         self.addParameter(sanitary_pref)
         search_distance = QgsProcessingParameterNumber(
             self.SEARCH_DISTANCE,
-            self.tr("Search distance"),
+            "Search distance",
             type=QgsProcessingParameterNumber.Double,
             defaultValue=10.0,
         )
@@ -161,13 +163,13 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         self.addParameter(search_distance)
 
     def processAlgorithm(self, parameters, context, feedback):
-        surface_lyr = self.parameterAsLayer(parameters, self.SURFACE_LAYER, context)
+        surface_lyr = self.parameterAsLayer(parameters, self.POLYGON_LAYER, context)
         if surface_lyr is None:
-            raise QgsProcessingException(self.invalidSourceError(parameters, self.SURFACE_LAYER))
-        surface_map_lyr = self.parameterAsLayer(parameters, self.SURFACE_MAP_LAYER, context)
+            raise QgsProcessingException(self.invalidSourceError(parameters, self.POLYGON_LAYER))
+        surface_map_lyr = self.parameterAsLayer(parameters, self.MAP_LAYER, context)
         if surface_map_lyr is None:
-            raise QgsProcessingException(self.invalidSourceError(parameters, self.SURFACE_MAP_LAYER))
-        selected_surfaces = self.parameterAsBool(parameters, self.SELECTED_SURFACES, context)
+            raise QgsProcessingException(self.invalidSourceError(parameters, self.MAP_LAYER))
+        selected_surfaces = self.parameterAsBool(parameters, self.SELECTED_POLYGONS, context)
         pipe_lyr = self.parameterAsLayer(parameters, self.PIPE_LAYER, context)
         if pipe_lyr is None:
             raise QgsProcessingException(self.invalidSourceError(parameters, self.PIPE_LAYER))
@@ -225,7 +227,6 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         surface_map_feats = []
         surface_map_fields = surface_map_lyr.fields()
         next_surface_map_id = get_next_feature_id(surface_map_lyr)
-        surface_id_field = "surface_id"
         for surface_id, surface_pipes in surface_to_pipes_distances.items():
             if feedback.isCanceled():
                 return {}
@@ -255,7 +256,7 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
                 node_geom = end_node_geom
             surface_map_feat = QgsFeature(surface_map_fields)
             surface_map_feat["id"] = next_surface_map_id
-            surface_map_feat[surface_id_field] = surface_feat["id"]
+            surface_map_feat[self.polygon_id_field] = surface_feat["id"]
             surface_map_feat["connection_node_id"] = surface_node_id
             surface_map_feat["percentage"] = 100.0
             surface_map_geom = QgsGeometry.fromPolylineXY([surface_centroid.asPoint(), node_geom.asPoint()])
@@ -276,3 +277,75 @@ class LinkSurfacesWithNodes(QgsProcessingAlgorithm):
         for layer in QgsProject.instance().mapLayers().values():
             layer.triggerRepaint()
         return {}
+
+
+class LinkSurfacesWithConnectionNodes(LinkToConnectionNodesAlgorithm):
+    """Link surfaces to connection nodes."""
+
+    @property
+    def parameter_names(self):
+        return {
+            LinkToConnectionNodesAlgorithm.POLYGON_LAYER: "Surface layer",
+            LinkToConnectionNodesAlgorithm.SELECTED_POLYGONS: "Selected surfaces only",
+            LinkToConnectionNodesAlgorithm.MAP_LAYER: "Surface map layer"
+        }
+
+    @property
+    def default_values(self):
+        return {
+            LinkToConnectionNodesAlgorithm.POLYGON_LAYER: "Surface",
+            LinkToConnectionNodesAlgorithm.MAP_LAYER: "Surface map",
+            LinkToConnectionNodesAlgorithm.SEWERAGE_TYPES: [
+                SewerageType.COMBINED_SEWER.value,
+                SewerageType.STORM_DRAIN.value
+            ]
+        }
+
+    @property
+    def polygon_id_field(self):
+        return "surface_id"
+
+    def createInstance(self):
+        return LinkSurfacesWithConnectionNodes()
+
+    def name(self):
+        return "threedi_map_surfaces_to_connection_nodes"
+
+    def displayName(self):
+        return "Map surfaces to connection nodes"
+
+
+class LinkDWFWithConnectionNodes(LinkToConnectionNodesAlgorithm):
+    """Link dry weather flow polygons to connection nodes."""
+
+    @property
+    def parameter_names(self):
+        return {
+            LinkToConnectionNodesAlgorithm.POLYGON_LAYER: "Dry weather flow layer",
+            LinkToConnectionNodesAlgorithm.SELECTED_POLYGONS: "Selected DWF polygons only",
+            LinkToConnectionNodesAlgorithm.MAP_LAYER: "Dry weather flow map layer"
+        }
+
+    @property
+    def default_values(self):
+        return {
+            LinkToConnectionNodesAlgorithm.POLYGON_LAYER: "Dry weather flow",
+            LinkToConnectionNodesAlgorithm.MAP_LAYER: "Dry weather flow map",
+            LinkToConnectionNodesAlgorithm.SEWERAGE_TYPES: [
+                SewerageType.COMBINED_SEWER.value,
+                SewerageType.SANITARY_SEWER.value
+            ]
+        }
+
+    @property
+    def polygon_id_field(self):
+        return "dry_weather_flow_id"
+
+    def createInstance(self):
+        return LinkDWFWithConnectionNodes()
+
+    def name(self):
+        return "threedi_map_dwf_to_connection_nodes"
+
+    def displayName(self):
+        return "Map dry weather flow to connection nodes"