+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 
 - Display upgrade warnings (any UserWarning raised during upgrading)
 - Added backwards compatibility to load a schematisation in 3.28. (#351)
+- Split BoundaryTypes in 1D and 2D types and add boundary types Discharge (total) and Groundwater discharge (total) to BoundaryType2D
 
 
 2.1.1 (2025-04-03)